+++ ehitchhiking-api/src/main/java/com/exadel/DBController/DAO/Dao.java
@@ -1,11 +1,10 @@
 package com.exadel.DBController.DAO;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface Dao<T> {
 
-    T get(long id);
+    T get(int id);
 
     List<T> getAll();
