+++ ehitchhiking-api/src/main/java/com/exadel/DBController/DAO/Trip_DriverDao.java
@@ -1,6 +1,5 @@
 package com.exadel.DBController.DAO;
 
-import com.exadel.DBController.Models.Driver;
 import com.exadel.DBController.Models.Trip_Driver;
 import com.exadel.DBController.Utils.HibernateSessionFactoryUtil;
 import org.hibernate.Session;
@@ -33,7 +32,7 @@ public class Trip_DriverDao implements Dao<Trip_Driver> {
         session.close();
     }
 
-    public Trip_Driver get(long id) {
+    public Trip_Driver get(int id) {
         return HibernateSessionFactoryUtil.getSessionFactory().openSession().get(Trip_Driver.class, id);
     }
